[WIP] ref bvh and hittable
